chore: rename to RouterTabs
